Code change to avoid signed/unsigned mismatch warnings.
This makes pdfium code on Linux and Mac sign-compare warning free.
The warning flag will be re-enabled after checking on windows clang build.

BUG=pdfium:29
R=tsepez@chromium.org

Review URL: https://codereview.chromium.org/1841643002 .

+++ core/fpdfapi/fpdf_page/cpdf_shadingpattern.cpp
@@ -43,7 +43,7 @@ CPDF_ShadingPattern::CPDF_ShadingPattern(CPDF_Document* pDoc,
     if (parentMatrix)
       m_Pattern2Form.Concat(*parentMatrix);
   }
-  for (int i = 0; i < FX_ArraySize(m_pFunctions); ++i)
+  for (size_t i = 0; i < FX_ArraySize(m_pFunctions); ++i)
     m_pFunctions[i] = nullptr;
 }
